+++ crates/derive/src/types/batch/span_batch/tx_data/eip2930.rs
@@ -37,7 +37,7 @@ impl SpanBatchEip2930TransactionData {
                     SpanBatchError::Decoding(SpanDecodingError::InvalidTransactionData)
                 })?,
             ),
-            gas_limit: gas,
+            gas_limit: gas as u128,
             to: if let Some(to) = to { TxKind::Call(to) } else { TxKind::Create },
             value: self.value,
             input: self.data.clone().into(),
